fixed and cleaned component editors
