+++ modular_framework_api/src/modular_framework_api/config_widgets/interface_config_widgets.py
@@ -16,7 +16,7 @@
 
 from PyQt5.QtWidgets import QWidget, QGridLayout, QLabel, QSpinBox, QHBoxLayout, QCheckBox
 from PyQt5.QtCore import pyqtSignal
-from editor_widgets import XMLEditorWidget
+from plain_editor_widgets import XMLEditorWidget
 import user_entry_widgets as uew
 import os
 import re
